requirements.txt
google
ipython
readability-lxml
html2text
werkzeug
pyyaml
pymongo
